backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
pydantic-settings==2.1.0

# Authentication & Security
PyJWT==2.10.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# Data Validation
pydantic==2.5.0
pydantic[email]==2.5.0

# HTTP & API
httpx==0.25.1
aiohttp==3.9.1
requests==2.31.0

# Vector Database & Caching
qdrant-client==2.7.2
redis==5.0.1

# LLM & AI
google-generativeai==0.5.0
google-ai-generativelanguage==0.4.0
langchain==0.1.0
langchain-community==0.0.8
langchain-google-genai==0.0.7
google-api-core==2.15.0
google-auth==2.27.0
google-cloud-core==2.4.1
grpcio==1.60.0
protobuf==4.25.1

# Telemetry & Observability
opik==0.2.10
google-cloud-logging==3.8.0
google-cloud-trace==1.11.1

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
isort==5.13.2
mypy==1.7.1

# Logging
structlog==23.2.0